requirements.txt
# Flask web framework
Flask==2.0.3

# Flask extensions
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0

# Database connector
psycopg2-binary==2.9.3
mysql-connector-python==8.4.0

# Testing libraries
pytest==7.1.2
pytest-cov==3.0.0

# Linting libraries
pylint==2.13.0

# Other useful libraries
requests==2.27.1
